Create modules and move methods to cut line count
