+++ lib/modules/season_stats.rb
@@ -1,0 +1,3 @@
+module SeasonStats
+  
+end
