+++ lib/modules/team_stats.rb
@@ -1,0 +1,3 @@
+module TeamStats
+
+end
