+++ test/stat_tracker_test.rb
@@ -172,10 +172,6 @@ class StatTrackerTest < Minitest::Test
   end
 
   def test_stat_tracker_can_get_worst_fans
-    @stat_tracker.worst_fans.each do |team|
-      assert_instance_of String, team
-    end
-
     assert_equal ['Houston Dynamo', 'Utah Royals FC'], @stat_tracker.worst_fans
   end
 end
